server/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
paho-mqtt==2.1.0
